Basic game is done. Implement pause, reset hi score, different game over screens.
So far, the basic game is done. The player is locked to his column while
the bullet is live. The score is updated only whenever it changes.
However, the conversion from unsigned char to char* has issues with
multiples of 10. There is a rudimentary reset system though it does not
seem to work all the time or reset all SM's.
TODO: Pause Menu -> Reset/View High Score
TODO: Check the EEPROM, not working.